Add module-level version of constant-propagation pass.
Summary:
Generally most of our passes should work on module level rather than on a graph
level, even though it might make sense for them to have graph version as well.
Constant propagation is a pretty popular pass that can be useful in many
scenarious as just a cleanup, and for it it definitely makes sense to have a
module-level version.

ghstack-source-id: 9aec897225f63c5824d320e55350cfddb173a1dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24285